+++ examples/resources/azexecute_application/resource.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     azexecute = {
       source  = "dyntora/azexecute"
-      version = "~> 0.6"
+      version = "~> 0.7"
     }
   }
 }